module12 hw3 completed
