+++ module12/hw3/suite_12_3.py
@@ -20,4 +20,17 @@
 Запустите TestSuite и проверьте полученные результаты тестов из обоих TestCase
 """
 import unittest
-from module12.hw3.tests_12_3 import TournamentTest
+from tests_12_3 import TournamentTest, RunnerTest
+
+
+# Создание тестового набора
+def suite():
+    suite = unittest.TestSuite()
+    suite.addTest(unittest.makeSuite(RunnerTest))
+    suite.addTest(unittest.makeSuite(TournamentTest))
+    return suite
+
+# Запуск тестов
+if __name__ == '__main__':
+    runner = unittest.TextTestRunner(verbosity=2)
+    runner.run(suite())
